pyproject.toml
[tool.poetry]
name = "e-image-editer"
version = "0.1.0"
description = ""
authors = ["ElliNet13 <ellinet13@googlegroups.com>"]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.12"
numpy = "^1.26.4"
pillow = "^10.3.0"

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"